pyproject.toml
[project]
name = "pdf-qa-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.7",
    "langchain>=0.3.24",
    "langchain-community>=0.3.23",
    "langchain-google-genai>=2.1.4",
    "marker-pdf[full]>=1.6.2",
    "pillow>=10.4.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.0",
]
